My addititons to the project
I made two major changes. First, i saw that you were trying to get a
gradient type thing to work on your moving line, so i went ahead and the
moving line section to make an actual gradient on the background that
changes every 250 frames. Second, when the mouse was clicked, it spawned
a random rectangle, but if it was placed too high, the program would
crash, so i went ahead and made it so that the rectangles would instead
be placed as high as possible, without crashing the program. I ran out
of time, but i was also going to make some additions to blur, as you
said you had trouble with that. Basically i was going to just locate
each pixel, add up the rgb values of the surrounding pixels, store this
as a local variable, divide by nine, and then apply those local rgb
values to all the pixels added up in that iteration.

Hope this helps,
-Koz